# apps/docs/static/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/households/{householdId}": {
      "get": {
        "operationId": "HouseholdsController_get_v1",
        "parameters": [
          {
            "name": "householdId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HouseholdDto"
                }
              }
            }
          },
          "default": {
            "description": "Erreur (code stable dans `error.code`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Détail d'un foyer et de ses membres",
        "tags": [
          "households"
        ]
      }
    },
...
    "/v1/households/{householdId}/activity": {
      "get": {
        "operationId": "ActivityController_list_v1",
        "parameters": [
          {
            "name": "householdId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "before",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 50,
              "type": "integer",
              "minimum": 1,
              "maximum": 500
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityPageDto"
                }
              }
            }
          },
          "default": {
            "description": "Erreur (code stable dans `error.code`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Journal d'activité, du plus récent au plus ancien (pagination par curseur)",
        "tags": [
          "activity"
        ]
      }
    },
    "/v1/households/{householdId}/trash": {
      "get": {
        "operationId": "ActivityController_trash_v1",
        "parameters": [
          {
            "name": "householdId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrashItemDto"
                  }
                }
              }
            }
          },
          "default": {
            "description": "Erreur (code stable dans `error.code`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Corbeille (éléments supprimés depuis moins de 30 jours)",
        "tags": [
          "activity"
        ]
      }
    },
    "/v1/households/{householdId}/trash/{id}/restore": {
      "post": {
        "operationId": "ActivityController_restore_v1",
        "parameters": [
          {
            "name": "householdId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-requested-with",
            "in": "header",
            "description": "Protection CSRF : valeur fixe.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "agenda-gn"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Succès"
          },
          "default": {
            "description": "Erreur (code stable dans `error.code`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Restaurer un élément de la corbeille",
        "tags": [
          "activity"
        ]
      }
    },
    "/v1/households/{householdId}/occurrences/{occurrenceId}/restore": {
      "post": {
        "operationId": "ActivityController_restoreOccurrence_v1",
        "parameters": [
          {
            "name": "householdId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "occurrenceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-requested-with",
            "in": "header",
            "description": "Protection CSRF : valeur fixe.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "agenda-gn"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Succès"
          },
          "default": {
            "description": "Erreur (code stable dans `error.code`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Annuler une suppression depuis la tâche elle-même (bouton « Annuler »)",
        "tags": [
          "activity"
        ]
      }
    },
...
    "/v1/households/{householdId}/absences": {
      "get": {
        "operationId": "AbsencesController_list_v1",
        "parameters": [
          {
            "name": "householdId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AbsenceDto"
                  }
                }
              }
            }
          },
          "default": {
            "description": "Erreur (code stable dans `error.code`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Absences en cours et à venir",
        "tags": [
          "absences"
        ]
      },
      "post": {
        "operationId": "AbsencesController_create_v1",
        "parameters": [
          {
            "name": "householdId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "x-requested-with",
            "in": "header",
            "description": "Protection CSRF : valeur fixe.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "agenda-gn"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAbsenceInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AbsenceDto"
                }
              }
            }
          },
          "default": {
            "description": "Erreur (code stable dans `error.code`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Déclarer une absence (les tâches partagées passent à l'autre membre)",
        "tags": [
          "absences"
        ]
      }
    },
    "/v1/households/{householdId}/absences/{id}": {
      "delete": {
        "operationId": "AbsencesController_remove_v1",
        "parameters": [
          {
            "name": "householdId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-requested-with",
            "in": "header",
            "description": "Protection CSRF : valeur fixe.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "agenda-gn"
              ]
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Aucun contenu"
          },
          "default": {
            "description": "Erreur (code stable dans `error.code`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Supprimer une absence",
        "tags": [
          "absences"
        ]
      }
    },
...
    "/v1/me/push-tokens/status": {
      "get": {
        "operationId": "PushController_status_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PushStatusDto"
                }
              }
            }
          },
          "default": {
            "description": "Erreur (code stable dans `error.code`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Diagnostic affiché dans les Réglages (site et app)",
        "tags": [
          "notifications"
        ]
      }
    },
    "/v1/me/push-tokens": {
      "put": {
        "operationId": "PushController_register_v1",
        "parameters": [
          {
            "name": "x-requested-with",
            "in": "header",
            "description": "Protection CSRF : valeur fixe.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "agenda-gn"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PushTokenInput"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Aucun contenu"
          },
          "default": {
            "description": "Erreur (code stable dans `error.code`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Enregistrer le jeton FCM de l'appareil Android",
        "tags": [
          "notifications"
        ]
      }
    },
    "/v1/me/push-tokens/{token}": {
      "delete": {
        "description": "Déconnexion : ce téléphone ne reçoit plus rien pour ce compte.",
        "operationId": "PushController_unregister_v1",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-requested-with",
            "in": "header",
            "description": "Protection CSRF : valeur fixe.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "agenda-gn"
              ]
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Aucun contenu"
          },
          "default": {
            "description": "Erreur (code stable dans `error.code`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Déconnexion ",
        "tags": [
          "notifications"
        ]
      }
    },
...
    "/v1/status": {
      "get": {
        "operationId": "StatusController_status_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusPageDto"
                }
              }
            }
          },
          "default": {
            "description": "Erreur (code stable dans `error.code`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "summary": "État public du service (page /status)",
        "tags": [
          "monitoring"
        ],
        "security": []
      }
    },
...
  },
  "info": {
    "title": "Agenda G & N — API",
    "description": "API REST de l'application **Agenda G & N** (web et Android). Toutes les routes sont versionnées\nsous `/v1`. Guide complet (authentification, erreurs, idempotence, temps réel) : page\n*API → Guide* de la documentation.\n\n- **Authentification** : `Authorization: Bearer <accessToken>` (Android, en-tête\n  `X-Client: mobile` à la connexion pour recevoir les jetons dans le corps), ou cookies\n  `httpOnly` posés par `/v1/auth/login` (web).\n- **CSRF** : toute requête qui modifie (POST, PATCH, PUT, DELETE) doit porter\n  `x-requested-with: agenda-gn`.\n- **Erreurs** : toujours `{ \"error\": { \"code\", \"message\", \"details\"? } }` ; le `code` est\n  stable (voir le schéma `ApiError`), le `message` n'est destiné qu'aux développeurs.\n- **Idempotence** : les créations de tâches acceptent `Idempotency-Key` (UUID) pour être\n  rejouées sans doublon depuis une file hors ligne.",
    "version": "1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/",
      "description": "Même origine que le site (le web relaie /v1/* vers l’API)"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "gn_at"
      }
    },
    "schemas": {
      "ApiError": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "VALIDATION_FAILED",
                  "UNAUTHENTICATED",
                  "INVALID_CREDENTIALS",
                  "EMAIL_ALREADY_USED",
                  "REGISTRATION_CLOSED",
                  "ACCOUNT_DISABLED",
                  "SESSION_EXPIRED",
                  "CSRF_REJECTED",
                  "FORBIDDEN",
                  "NOT_FOUND",
                  "HOUSEHOLD_NOT_FOUND",
                  "INVITATION_INVALID",
                  "RESET_TOKEN_INVALID",
                  "GOOGLE_NOT_CONFIGURED",
                  "GOOGLE_FAILED",
                  "GOOGLE_EMAIL_EXISTS",
                  "GOOGLE_ALREADY_LINKED",
                  "CURRENT_PASSWORD_INVALID",
                  "PASSWORD_REQUIRED",
                  "CONFIRMATION_REQUIRED",
                  "CALENDAR_NOT_CONFIGURED",
                  "CALENDAR_NOT_CONNECTED",
                  "CALENDAR_READ_ONLY",
                  "CALENDAR_NOT_FOUND",
                  "GOOGLE_REVOKED",
                  "GOOGLE_FLOW_EXPIRED",
                  "GOOGLE_DENIED",
                  "GOOGLE_SCOPES_MISSING",
                  "ALREADY_MEMBER",
                  "VERSION_CONFLICT",
                  "IDEMPOTENCY_KEY_INVALID",
                  "IDEMPOTENCY_KEY_REUSED",
                  "IDEMPOTENCY_IN_PROGRESS",
                  "CATEGORY_NAME_TAKEN",
                  "TASK_TITLE_REQUIRED",
                  "RATE_LIMITED",
                  "ATTACHMENT_TOO_LARGE",
                  "ATTACHMENTS_QUOTA",
                  "INTERNAL"
                ]
              },
              "message": {
                "type": "string",
                "description": "Message technique (anglais), jamais affiché."
              },
              "details": {
                "description": "Détail éventuel (champs invalides pour VALIDATION_FAILED)."
              }
            }
          }
        }
      },
      "AbsenceDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "memberId": {
            "type": "string",
            "format": "uuid"
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "format": "date"
          }
        },
        "required": [
          "id",
          "memberId",
          "startDate",
          "endDate"
        ],
        "additionalProperties": false
      },
...
      "ActivityPageDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "action": {
                  "type": "string"
                },
                "at": {
                  "type": "string"
                },
                "actorId": {
                  "nullable": true,
                  "type": "string",
                  "format": "uuid"
                },
                "title": {
                  "nullable": true,
                  "type": "string"
                },
                "date": {
                  "nullable": true,
                  "type": "string",
                  "format": "date"
                },
                "fields": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              },
              "required": [
                "id",
                "action",
                "at",
                "actorId",
                "title",
                "date",
                "fields"
              ],
              "additionalProperties": false
            }
          },
          "next": {
            "nullable": true,
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "items",
          "next"
        ],
        "additionalProperties": false
      },
...
      "CreateAbsenceInput": {
        "type": "object",
        "properties": {
          "memberId": {
            "type": "string",
            "format": "uuid"
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "format": "date"
          }
        },
        "required": [
          "memberId",
          "startDate",
          "endDate"
        ]
      },
...
      "HouseholdDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "timezone": {
            "type": "string"
          },
          "members": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "userId": {
                  "nullable": true,
                  "type": "string",
                  "format": "uuid"
                },
                "displayName": {
                  "type": "string"
                },
                "role": {
                  "type": "string",
                  "enum": [
                    "OWNER",
                    "MEMBER"
                  ]
                },
                "color": {
                  "type": "string",
                  "enum": [
                    "sage",
                    "ocean",
                    "amber",
                    "plum",
                    "clay",
                    "slate"
                  ]
                },
                "absentUntil": {
                  "nullable": true,
                  "type": "string"
                }
              },
              "required": [
                "id",
                "userId",
                "displayName",
                "role",
                "color"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "id",
          "name",
          "timezone",
          "members"
        ],
        "additionalProperties": false
      },
...
      "PushStatusDto": {
        "type": "object",
        "properties": {
          "serverEnabled": {
            "type": "boolean"
          },
          "serverIssue": {
            "nullable": true,
            "type": "string",
            "enum": [
              "NOT_CONFIGURED",
              "INVALID_CONFIG"
            ]
          },
          "devices": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "lastRegisteredAt": {
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "serverEnabled",
          "serverIssue",
          "devices",
          "lastRegisteredAt"
        ],
        "additionalProperties": false
      },
      "PushTokenInput": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "minLength": 10,
            "maxLength": 4096
          },
          "platform": {
            "default": "android",
            "type": "string",
            "enum": [
              "android"
            ]
          }
        },
        "required": [
          "token"
        ]
      },
...
      "StatusPageDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "operational",
              "degraded",
              "down",
              "unknown"
            ]
          },
          "updatedAt": {
            "type": "string"
          },
          "components": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "key": {
                  "type": "string",
                  "enum": [
                    "api",
                    "database",
                    "redis",
                    "calendar",
                    "email",
                    "backups"
                  ]
                },
                "status": {
                  "type": "string",
                  "enum": [
                    "operational",
                    "degraded",
                    "down",
                    "unknown"
                  ]
                },
                "latencyMs": {
                  "nullable": true,
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991
                },
                "uptime90": {
                  "nullable": true,
                  "type": "number"
                },
                "days": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "date": {
                        "type": "string"
                      },
                      "uptime": {
                        "nullable": true,
                        "type": "number"
                      }
                    },
                    "required": [
                      "date",
                      "uptime"
                    ],
                    "additionalProperties": false
                  }
                }
              },
              "required": [
                "key",
                "status",
                "latencyMs",
                "uptime90",
                "days"
              ],
              "additionalProperties": false
            }
          },
          "incidents": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "component": {
                  "type": "string",
                  "enum": [
                    "api",
                    "database",
                    "redis",
                    "calendar",
                    "email",
                    "backups"
                  ]
                },
                "startedAt": {
                  "type": "string"
                },
                "resolvedAt": {
                  "nullable": true,
                  "type": "string"
                }
              },
              "required": [
                "id",
                "component",
                "startedAt",
                "resolvedAt"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "status",
          "updatedAt",
          "components",
          "incidents"
        ],
        "additionalProperties": false
      },
...
      "TrashItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "kind": {
            "type": "string",
            "enum": [
              "task",
              "occurrence",
              "following"
            ]
          },
          "title": {
            "type": "string"
          },
          "date": {
            "nullable": true,
            "type": "string",
            "format": "date"
          },
          "deletedAt": {
            "type": "string"
          },
          "deletedById": {
            "nullable": true,
            "type": "string",
            "format": "uuid"
          },
          "purgeAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "title",
          "date",
          "deletedAt",
          "deletedById",
          "purgeAt"
        ],
        "additionalProperties": false
      },
...
    }
  }
}